# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: InsuraBridge - Insurance Middleware Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List all claims
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          schema:
            type: string
        - name: hospitalId
          in: query
          required: false
          schema:
            type: string
        - name: tpaId
          in: query
          required: false
          schema:
            type: string
        - name: insurerId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
    post:
      operationId: createClaim
      tags: [claims]
      summary: Create a new claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimRequest"
      responses:
        "201":
          description: Claim created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /claims/{id}:
    get:
      operationId: getClaim
      tags: [claims]
      summary: Get claim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claim found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
    patch:
      operationId: updateClaim
      tags: [claims]
      summary: Update claim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClaimRequest"
      responses:
        "200":
          description: Claim updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"

  # MESSAGES & THREADS
...
  /ecards:
    get:
      operationId: listEcards
      tags: [ecards]
      summary: List e-cards
      parameters:
        - name: memberId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of e-cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ecard"
    post:
      operationId: createEcard
      tags: [ecards]
      summary: Issue a new e-card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEcardRequest"
      responses:
        "201":
          description: E-card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ecard"
  /ecards/{id}:
    patch:
      operationId: updateEcard
      tags: [ecards]
      summary: Update e-card status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEcardRequest"
      responses:
        "200":
          description: E-card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ecard"

  # NETWORK PROVIDERS (TPA feature)
  /network-providers:
    get:
      operationId: listNetworkProviders
      tags: [network]
      summary: List network hospitals/providers
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: speciality
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of network providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NetworkProvider"
    post:
      operationId: createNetworkProvider
      tags: [network]
      summary: Add a network provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNetworkProviderRequest"
      responses:
        "201":
          description: Network provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkProvider"

  # SCRUTINY SETTLEMENTS (TPA feature)
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List policy members
      parameters:
        - name: policyNumber
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: addMember
      tags: [members]
      summary: Add a new member to a policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # REIMBURSEMENT SETTLEMENTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, hospital, tpa, insurer, admin]
        organization:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
      required: [id, name, email, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
...
    Claim:
      type: object
      properties:
        id:
          type: integer
        claimNumber:
          type: string
        customerId:
          type: integer
        patientName:
          type: string
        hospitalId:
          type: integer
          nullable: true
        hospitalName:
          type: string
          nullable: true
        insurerId:
          type: integer
          nullable: true
        tpaId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [submitted, under_review, approved, rejected, settled, pending_documents]
        claimType:
          type: string
          enum: [cashless, reimbursement]
        diagnosis:
          type: string
          nullable: true
        admissionDate:
          type: string
          format: date-time
          nullable: true
        dischargeDate:
          type: string
          format: date-time
          nullable: true
        claimedAmount:
          type: number
        approvedAmount:
          type: number
          nullable: true
        deductible:
          type: number
          nullable: true
        coPayAmount:
          type: number
          nullable: true
        netPayableAmount:
          type: number
          nullable: true
        roomRentCharges:
          type: number
          nullable: true
        surgeryCharges:
          type: number
          nullable: true
        medicineCharges:
          type: number
          nullable: true
        diagnosticCharges:
          type: number
          nullable: true
        otherCharges:
          type: number
          nullable: true
        policyNumber:
          type: string
        icdCode:
          type: string
          nullable: true
        treatmentType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - claimNumber
        - customerId
        - patientName
        - status
        - claimType
        - claimedAmount
        - policyNumber
        - createdAt
        - updatedAt
    CreateClaimRequest:
      type: object
      properties:
        customerId:
          type: integer
        patientName:
          type: string
        hospitalId:
          type: integer
          nullable: true
        hospitalName:
          type: string
          nullable: true
        insurerId:
          type: integer
          nullable: true
        tpaId:
          type: integer
          nullable: true
        claimType:
          type: string
          enum: [cashless, reimbursement]
        diagnosis:
          type: string
          nullable: true
        admissionDate:
          type: string
          format: date-time
          nullable: true
        dischargeDate:
          type: string
          format: date-time
          nullable: true
        claimedAmount:
          type: number
        roomRentCharges:
          type: number
          nullable: true
        surgeryCharges:
          type: number
          nullable: true
        medicineCharges:
          type: number
          nullable: true
        diagnosticCharges:
          type: number
          nullable: true
        otherCharges:
          type: number
          nullable: true
        policyNumber:
          type: string
        icdCode:
          type: string
          nullable: true
        treatmentType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [customerId, patientName, claimType, claimedAmount, policyNumber]
    UpdateClaimRequest:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, under_review, approved, rejected, settled, pending_documents]
        approvedAmount:
          type: number
          nullable: true
        deductible:
          type: number
          nullable: true
        coPayAmount:
          type: number
          nullable: true
        netPayableAmount:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        diagnosis:
          type: string
          nullable: true
...
    Ecard:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        policyNumber:
          type: string
        insurerName:
          type: string
        tpaName:
          type: string
        cardNumber:
          type: string
        sumInsured:
          type: number
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, expired, suspended, cancelled]
        dependents:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, memberId, memberName, policyNumber, insurerName, tpaName, cardNumber, sumInsured, validFrom, validTo, status, dependents, createdAt]
    CreateEcardRequest:
      type: object
      properties:
        memberId:
          type: integer
        memberName:
          type: string
        policyNumber:
          type: string
        insurerName:
          type: string
        tpaName:
          type: string
        sumInsured:
          type: number
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        dependents:
          type: array
          items:
            type: string
      required: [memberId, memberName, policyNumber, insurerName, tpaName, sumInsured, validFrom, validTo]
    UpdateEcardRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, expired, suspended, cancelled]
        sumInsured:
          type: number
          nullable: true
        validTo:
          type: string
          format: date-time
          nullable: true
    NetworkProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [hospital, clinic, diagnostic, pharmacy, specialist]
        city:
          type: string
        state:
          type: string
        address:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        specialities:
          type: array
          items:
            type: string
        insurerIds:
          type: array
          items:
            type: integer
        bedCount:
          type: integer
          nullable: true
        cashlessEnabled:
          type: boolean
        status:
          type: string
          enum: [active, inactive, suspended]
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, city, state, address, specialities, insurerIds, cashlessEnabled, status, createdAt]
    CreateNetworkProviderRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [hospital, clinic, diagnostic, pharmacy, specialist]
        city:
          type: string
        state:
          type: string
        address:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        specialities:
          type: array
          items:
            type: string
        insurerIds:
          type: array
          items:
            type: integer
        bedCount:
          type: integer
          nullable: true
        cashlessEnabled:
          type: boolean
      required: [name, type, city, state, address, specialities, insurerIds, cashlessEnabled]
...
    Member:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        policyNumber:
          type: string
        name:
          type: string
        relationship:
          type: string
          enum: [self, spouse, child, parent, parent_in_law, sibling]
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
          enum: [male, female, other]
        preExistingConditions:
          type: array
          items:
            type: string
        sumInsured:
          type: number
        status:
          type: string
          enum: [active, inactive]
        addedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, customerId, policyNumber, name, relationship, dateOfBirth, gender, preExistingConditions, sumInsured, status, addedAt, createdAt]
    AddMemberRequest:
      type: object
      properties:
        customerId:
          type: integer
        policyNumber:
          type: string
        name:
          type: string
        relationship:
          type: string
          enum: [self, spouse, child, parent, parent_in_law, sibling]
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
          enum: [male, female, other]
        preExistingConditions:
          type: array
          items:
            type: string
        sumInsured:
          type: number
        addedAt:
          type: string
          format: date-time
      required: [customerId, policyNumber, name, relationship, dateOfBirth, gender, sumInsured, addedAt]
...